+++ src/main/resources/templates/member/fragment/membersidebar.html
@@ -24,10 +24,10 @@
                 </a>
             </li>
 
-            <!-- 나의 구매내역 -->
+            <!-- 나의 주문내역 -->
             <li class="nav-item my-2">
                 <a href="/myOrders" class="nav-link text-white sidebar-item" data-url="myOrders">
-                    <i class="fas fa-boxes me-2"></i>나의 구매내역
+                    <i class="fas fa-boxes me-2"></i>나의 주문내역
                 </a>
             </li>
 
@@ -40,9 +40,16 @@
 
             <hr class="sidebar-divider">
 
-            <!-- 로그 아웃 -->
+            <!-- 회원 탈퇴 -->
             <li class="nav-item my-2">
-                <a href="#" class="nav-link text-white sidebar-item" data-url="/logout">
+                <a href="/" class="nav-link text-white sidebar-item" data-url="deleteAccount">
+                    <i class="fas fa-user-times me-2"></i>회원 탈퇴
+                </a>
+            </li>
+
+            <!-- 로그아웃 -->
+            <li class="nav-item my-2">
+                <a href="#" class="nav-link text-white sidebar-item" data-url="logout">
                     <i class="fas fa-sign-out-alt me-2"></i>로그아웃
                 </a>
             </li>
