+++ src/main/resources/templates/member/myCart.html
@@ -6,7 +6,8 @@
 </head>
 <body>
 <div layout:fragment="content">
-
+    <h1>나의 장바구니</h1>
+    <p>나의 장바구니</p>
 </div>
 </body>
 </html>
