+++ src/main/resources/templates/member/myCoupons.html
@@ -6,7 +6,8 @@
 </head>
 <body>
 <div layout:fragment="content">
-
+    <h1>나의 쿠폰</h1>
+    <p>나의 쿠폰</p>
 </div>
 </body>
 </html>
