+++ src/main/resources/templates/member/myOrders.html
@@ -6,7 +6,8 @@
 </head>
 <body>
 <div layout:fragment="content">
-
+    <h1>나의 주문내역</h1>
+    <p>나의 주문내역</p>
 </div>
 </body>
 </html>
